Move `Project` into `workspace` module (#364)
* Move `Project` into `workspace` module

* private
